# Conflicts:
#	PRG2_T06_Team9/PRG2_T06_Team9/obj/Debug/netcoreapp3.1/PRG2_T06_Team9.csprojAssemblyReference.cache

+++ PRG2_T06_Team9/PRG2_T06_Team9/Visitor.cs
@@ -29,7 +29,7 @@ namespace PRG2_T06_Team9
 =======
             //Checking for the last country of embarkation
 
-            float price = 0;
+            double price = 0;
             if (person.LastCountryOfEmbarkation == "New Zealand" && person.LastCountryOfEmbarkation == "Vietnam")
             {
                 if (person.CalculateSHNDuration() == 0)
